[TASK] Round down calculated maximum height
The SrcSet service needs to determine whether a responsive variant can be calculated for a given image.
If a ratio is defined, the maximum height is calculated with the following division: width / ratio
This will not always produce an integer, e.g. if width is 960px and ratio 2.1/1, the calculated height would
be 457.1428(...) pixels.
If the original image is 960*457 pixels, it would be smaller than the calculated height and therefore the
SrcSetService would assume that it cannot generate a variante with 960px width.
To solve this, we need to round down the calculated maximum height using floor().